Split the job_lookup namespace to avoid cyclical import. Let transcriptome jobs run.
