+++ common/data_refinery_common/rna_seq.py
@@ -2,7 +2,7 @@ from typing import List
 
 from django.db.models import OuterRef, Subquery
 
-from data_refinery_common.job_lookup import ProcessorEnum
+from data_refinery_common.enums import ProcessorEnum
 from data_refinery_common.logging import get_and_configure_logger
 from data_refinery_common.models import ComputationalResult, Experiment
 from data_refinery_common.utils import get_env_variable
