+++ foreman/data_refinery_foreman/foreman/job_control.py
@@ -4,7 +4,7 @@ import time
 from django.conf import settings
 from django.utils import timezone
 
-from data_refinery_common.job_lookup import ProcessorPipeline
+from data_refinery_common.enums import ProcessorPipeline
 from data_refinery_common.logging import get_and_configure_logger
 from data_refinery_common.message_queue import send_job
 from data_refinery_common.models import ComputedFile, ProcessorJob
