+++ foreman/data_refinery_foreman/foreman/job_requeuing.py
@@ -1,4 +1,4 @@
-from data_refinery_common.job_lookup import Downloaders, ProcessorPipeline, SurveyJobTypes
+from data_refinery_common.enums import Downloaders, ProcessorPipeline, SurveyJobTypes
 from data_refinery_common.logging import get_and_configure_logger
 from data_refinery_common.message_queue import send_job
 from data_refinery_common.models import (
