+++ foreman/data_refinery_foreman/foreman/management/commands/create_compendia.py
@@ -4,7 +4,7 @@ from django.core.management.base import BaseCommand
 from django.db.models.aggregates import Count
 from django.db.models.expressions import Q
 
-from data_refinery_common.job_lookup import ProcessorPipeline
+from data_refinery_common.enums import ProcessorPipeline
 from data_refinery_common.logging import get_and_configure_logger
 from data_refinery_common.message_queue import send_job
 from data_refinery_common.models import (
