+++ foreman/data_refinery_foreman/foreman/management/commands/rerun_salmon_old_samples.py
@@ -8,7 +8,7 @@ from django.core.management.base import BaseCommand
 from django.db.models import Count
 from django.db.models.expressions import Q
 
-from data_refinery_common.job_lookup import ProcessorEnum, ProcessorPipeline
+from data_refinery_common.enums import ProcessorEnum, ProcessorPipeline
 from data_refinery_common.job_management import create_downloader_job
 from data_refinery_common.logging import get_and_configure_logger
 from data_refinery_common.models import Experiment, ProcessorJob
