+++ foreman/data_refinery_foreman/foreman/management/commands/run_tximport.py
@@ -11,7 +11,7 @@ from typing import List
 from django.core.management.base import BaseCommand
 from django.db.models import Count
 
-from data_refinery_common.job_lookup import ProcessorPipeline
+from data_refinery_common.enums import ProcessorPipeline
 from data_refinery_common.logging import get_and_configure_logger
 from data_refinery_common.message_queue import send_job
 from data_refinery_common.models import (
