+++ foreman/data_refinery_foreman/foreman/management/commands/test_run_tximport.py
@@ -4,7 +4,7 @@ from unittest.mock import patch
 
 from django.test import TestCase
 
-from data_refinery_common.job_lookup import ProcessorPipeline
+from data_refinery_common.enums import ProcessorPipeline
 from data_refinery_common.models import (
     ComputationalResult,
     ComputationalResultAnnotation,
