+++ foreman/data_refinery_foreman/foreman/processor_job_manager.py
@@ -1,7 +1,7 @@
 from typing import List
 
 import data_refinery_foreman.foreman.utils as utils
-from data_refinery_common.job_lookup import ProcessorPipeline
+from data_refinery_common.enums import ProcessorPipeline
 from data_refinery_common.logging import get_and_configure_logger
 from data_refinery_common.message_queue import get_capacity_for_jobs, send_job
 from data_refinery_common.models import ProcessorJob
