+++ foreman/data_refinery_foreman/surveyor/geo.py
@@ -5,7 +5,7 @@ from typing import Dict, List
 import dateutil.parser
 import GEOparse
 
-from data_refinery_common.job_lookup import Downloaders
+from data_refinery_common.enums import Downloaders
 from data_refinery_common.logging import get_and_configure_logger
 from data_refinery_common.models import (
     Experiment,
