+++ foreman/data_refinery_foreman/surveyor/sra.py
@@ -5,7 +5,7 @@ from typing import Dict, List
 
 from django.utils.dateparse import parse_date
 
-from data_refinery_common.job_lookup import Downloaders
+from data_refinery_common.enums import Downloaders
 from data_refinery_common.logging import get_and_configure_logger
 from data_refinery_common.models import (
     Experiment,
