+++ foreman/data_refinery_foreman/surveyor/test_transcriptome_index.py
@@ -5,7 +5,7 @@ from django.test import TestCase
 
 import vcr
 
-from data_refinery_common.job_lookup import Downloaders
+from data_refinery_common.enums import Downloaders
 from data_refinery_common.models import DownloaderJob, Organism, SurveyJob, SurveyJobKeyValue
 from data_refinery_foreman.surveyor.transcriptome_index import TranscriptomeIndexSurveyor
