+++ workers/data_refinery_workers/downloaders/transcriptome_index.py
@@ -3,7 +3,7 @@ import shutil
 import urllib.request
 from contextlib import closing
 
-from data_refinery_common.job_lookup import ProcessorPipeline
+from data_refinery_common.enums import ProcessorPipeline
 from data_refinery_common.logging import get_and_configure_logger
 from data_refinery_common.message_queue import send_job
 from data_refinery_common.models import (
