+++ workers/data_refinery_workers/processors/array_express.py
@@ -8,7 +8,7 @@ from django.utils import timezone
 import rpy2.robjects as ro
 from rpy2.rinterface import RRuntimeError
 
-from data_refinery_common.job_lookup import PipelineEnum
+from data_refinery_common.enums import PipelineEnum
 from data_refinery_common.logging import get_and_configure_logger
 from data_refinery_common.models import (
     ComputationalResult,
