+++ workers/data_refinery_workers/processors/create_quantpendia.py
@@ -9,7 +9,7 @@ from django.utils import timezone
 
 import psutil
 
-from data_refinery_common.job_lookup import PipelineEnum
+from data_refinery_common.enums import PipelineEnum
 from data_refinery_common.logging import get_and_configure_logger
 from data_refinery_common.models import (
     CompendiumResult,
