+++ workers/data_refinery_workers/processors/no_op.py
@@ -3,7 +3,7 @@ import os
 import subprocess
 from typing import Dict
 
-from data_refinery_common.job_lookup import PipelineEnum
+from data_refinery_common.enums import PipelineEnum
 from data_refinery_common.logging import get_and_configure_logger
 from data_refinery_common.models import (
     ComputationalResult,
