+++ workers/data_refinery_workers/processors/qn_reference.py
@@ -6,7 +6,7 @@ from django.utils import timezone
 
 import pandas as pd
 
-from data_refinery_common.job_lookup import PipelineEnum
+from data_refinery_common.enums import PipelineEnum
 from data_refinery_common.logging import get_and_configure_logger
 from data_refinery_common.models import (
     ComputationalResult,
