+++ workers/data_refinery_workers/processors/salmon.py
@@ -16,7 +16,7 @@ import pandas as pd
 import untangle
 from botocore.client import Config
 
-from data_refinery_common.job_lookup import PipelineEnum
+from data_refinery_common.enums import PipelineEnum
 from data_refinery_common.logging import get_and_configure_logger
 from data_refinery_common.models import (
     ComputationalResult,
