+++ workers/data_refinery_workers/processors/test_create_quantpendia.py
@@ -2,7 +2,7 @@ import os
 
 from django.test import TransactionTestCase, tag
 
-from data_refinery_common.job_lookup import ProcessorPipeline
+from data_refinery_common.enums import ProcessorPipeline
 from data_refinery_common.models import (
     ComputationalResult,
     ComputedFile,
