+++ workers/data_refinery_workers/processors/transcriptome_index.py
@@ -7,7 +7,7 @@ from typing import Dict
 
 from django.utils import timezone
 
-from data_refinery_common.job_lookup import PipelineEnum
+from data_refinery_common.enums import PipelineEnum
 from data_refinery_common.logging import get_and_configure_logger
 from data_refinery_common.models import (
     ComputationalResult,
